for the visit restore button in dataview/search
