+++ packages/datagateway-common/src/downloadConfirmation/DownloadRequestInfo.ts
