+++ packages/datagateway-common/src/downloadConfirmation/dialogContent.component.ts
