+++ packages/datagateway-download/src/downloadApiHooks.test.tsx
@@ -5,7 +5,7 @@ import {
 } from '@testing-library/react-hooks';
 import axios, { AxiosError } from 'axios';
 import { Download, InvalidateTokenType } from 'datagateway-common';
-import { handleICATError, NotificationType } from 'datagateway-common';
+import { handleICATError } from 'datagateway-common';
 import { createMemoryHistory } from 'history';
 import * as React from 'react';
 import { QueryClient, QueryClientProvider, setLogger } from 'react-query';
@@ -21,14 +21,12 @@ import {
   useDownloadOrRestoreDownload,
   useDownloadPercentageComplete,
   useDownloads,
-  useDownloadTypeStatuses,
   useFileSizesAndCounts,
   useIsCartMintable,
   useIsTwoLevel,
   useMintCart,
   useRemoveAllFromCart,
   useRemoveEntityFromCart,
-  useSubmitCart,
 } from './downloadApiHooks';
 import { mockCartItems, mockDownloadItems, mockedSettings } from './testData';
 import log from 'loglevel';
@@ -975,195 +973,6 @@ describe('Download API react-query hooks test', () => {
       expect(result.current.useAdminDownloads.data?.pages).toEqual([
         mockDownloadItems,
       ]);
-    });
-  });
-
-  describe('useSubmitCart', () => {
-    it('should submit cart and clear cart on success', async () => {
-      axios.post = jest.fn().mockResolvedValue({ data: 123 });
-      axios.get = jest
-        .fn()
-        .mockResolvedValueOnce({
-          data: {
-            cartItems: mockCartItems,
-          },
-        })
-        .mockResolvedValueOnce({ data: { cartItems: [] } });
-
-      const { result, waitFor } = renderHook(
-        () => ({
-          useSubmitCart: useSubmitCart(),
-          useCart: useCart(),
-        }),
-        { wrapper: createReactQueryWrapper() }
-      );
-
-      // wait for the cart to finish loading
-      await waitFor(() => result.current.useCart.isSuccess);
-      // submit the cart
-      result.current.useSubmitCart.mutate({
-        emailAddress: 'cat@dog.com',
-        fileName: 'test-file',
-        transport: 'https',
-      });
-      // wait for cart submission to finish
-      await waitFor(() => result.current.useSubmitCart.isSuccess);
-
-      expect(result.current.useCart.data).toEqual([]);
-    });
-
-    it('should call handleICATError when an error is encountered', async () => {
-      axios.post = jest.fn().mockRejectedValue({
-        message: 'test error message',
-      });
-      axios.get = jest.fn().mockResolvedValueOnce({
-        data: {
-          cartItems: mockCartItems,
-        },
-      });
-
-      const { result, waitFor } = renderHook(
-        () => ({
-          useSubmitCart: useSubmitCart(),
-          useCart: useCart(),
-        }),
-        { wrapper: createReactQueryWrapper() }
-      );
-
-      await waitFor(() => result.current.useCart.isSuccess);
-      result.current.useSubmitCart.mutate({
-        emailAddress: 'a@b.c',
-        fileName: 'test-file',
-        transport: 'https',
-      });
-      await waitFor(() => result.current.useSubmitCart.isError);
-
-      expect(handleICATError).toHaveBeenCalledWith({
-        message: 'test error message',
-      });
-    });
-  });
-
-  describe('useDownloadTypeStatuses', () => {
-    const downloadTypes = ['https', 'globus'];
-
-    let queryClient: QueryClient;
-
-    beforeAll(() => {
-      queryClient = new QueryClient();
-    });
-
-    afterEach(() => {
-      queryClient.clear();
-    });
-
-    it('should query statuses of download types', async () => {
-      axios.get = jest.fn().mockResolvedValue({
-        data: {
-          disabled: false,
-          message: '',
-        },
-      });
-
-      const { result, waitFor } = renderHook(
-        () => useDownloadTypeStatuses({ downloadTypes }),
-        { wrapper: createReactQueryWrapper() }
-      );
-
-      await waitFor(() => result.current.every((query) => query.isSuccess));
-
-      const data = result.current.map(({ data }) => data);
-      expect(data).toEqual([
-        {
-          type: 'https',
-          disabled: false,
-          message: '',
-        },
-        {
-          type: 'globus',
-          disabled: false,
-          message: '',
-        },
-      ]);
-    });
-
-    it('should dispatch event with the error messages of download type queries with errors', async () => {
-      axios.get = jest
-        .fn()
-        .mockResolvedValueOnce({
-          data: {
-            disabled: false,
-            message: '',
-          },
-        })
-        .mockImplementationOnce(() =>
-          Promise.reject({
-            message: 'Test error message',
-          })
-        );
-
-      const dispatchEventSpy = jest.spyOn(document, 'dispatchEvent');
-
-      const { result, waitFor } = renderHook(
-        () => useDownloadTypeStatuses({ downloadTypes }),
-        { wrapper: createReactQueryWrapper() }
-      );
-
-      await waitFor(() =>
-        result.current.every((query) => query.isSuccess || query.isError)
-      );
-
-      expect((dispatchEventSpy.mock.calls[0][0] as CustomEvent).detail).toEqual(
-        {
-          type: NotificationType,
-          payload: {
-            severity: 'error',
-            message:
-              'downloadConfirmDialog.access_method_error {method:GLOBUS}',
-          },
-        }
-      );
-    });
-
-    it('should refetch data on every hook call', async () => {
-      axios.get = jest.fn().mockResolvedValue({
-        data: {
-          disabled: false,
-          message: '',
-        },
-      });
-
-      const wrapper = createReactQueryWrapper();
-
-      const { result, waitFor } = renderHook(
-        () =>
-          useDownloadTypeStatuses({
-            downloadTypes: ['https'],
-          }),
-        { wrapper }
-      );
-
-      await waitFor(() => result.current.every((query) => query.isSuccess));
-
-      expect(result.current[0].isStale).toBe(true);
-      expect(axios.get).toHaveBeenCalledTimes(1);
-
-      await act(async () => {
-        const { result: newResult } = renderHook(
-          () =>
-            useDownloadTypeStatuses({
-              downloadTypes: ['https'],
-            }),
-          { wrapper }
-        );
-
-        await waitFor(() =>
-          newResult.current.every((query) => query.isSuccess)
-        );
-
-        expect(newResult.current[0].isStale).toBe(true);
-        expect(axios.get).toHaveBeenCalledTimes(2);
-      });
     });
   });
